second

+++ src/Server.cc
@@ -1,15 +1,16 @@
-#include "Server.h"
-#include <sys/socket.h>
-#include <sys/un.h>
+#include <iostream>
+
 #include <string.h>
 
-#include <iostream>
+#include "CommonDef.h"
+#include "Server.h"
 
 using namespace std;
 
-Server::Server(const string& domain)
+Server::Server(const string& domain, bool isFile)
 {
     domainPath = domain;
+    this->isFile = isFile;
     LoopStart();
 }
 
@@ -18,18 +19,24 @@ Server::~Server()
     LoopStop();
 }
 
-int Server::Open(const string& path)
+int Server::Open(const string& path, bool isFile)
 {
     const char* domain = path.c_str();
-    struct sockaddr_un server_address;
-    unlink(domain);
     int server_sockfd = ::socket(AF_UNIX, SOCK_SEQPACKET, 0);
-    cout << "path: " << domain << "fd: " << server_sockfd << endl;
+    struct sockaddr_un server_address;
     server_address.sun_family = AF_UNIX;
-    strcpy (server_address.sun_path, domain);
-    int server_len = sizeof (server_address);
-    cout << "bind:" << ::bind (server_sockfd, (struct sockaddr *)&server_address, server_len)<<endl;
-    cout << "listen:" << ::listen (server_sockfd, 5) << endl;
+    server_address.sun_path[0] = 0;
+    strncpy(server_address.sun_path+!isFile, domain, UNIX_PATH_MAX-1);
+    cout << "fd:" << server_sockfd << " isFile:" << isFile << " domain:" << domain << endl;
+    int address_len = sizeof(server_address.sun_family) + strlen(domain) + 1;
+    if(isFile)
+    {
+        address_len = sizeof(server_address);
+        unlink(domain);
+    }
+    int b=::bind (server_sockfd, (struct sockaddr *)&server_address, address_len);
+    int l=::listen (server_sockfd, 5);
+    cout << "b:" << b << " l:" << l << endl;
     return server_sockfd;
 }
 
@@ -42,7 +49,7 @@ void Server::Loop()
         struct sockaddr_un client_address;
         int client_len = sizeof(client_address);
         int client_sockfd = ::accept(server_sockfd, (struct sockaddr *)&client_address, (socklen_t *)&client_len);
-        cout << "new FD:" << client_sockfd << " errno: " << errno<<"-" <<strerror(errno) << endl;
+ //       cout << "new FD:" << client_sockfd << " errno: " << errno<<"-" <<strerror(errno) << endl;
         if(client_sockfd < 0)
             continue;
         slist.push_back(client_sockfd);
@@ -51,7 +58,7 @@ void Server::Loop()
 
 void Server::LoopStart()
 {
-    sockFd = Open(domainPath);
+    sockFd = Open(domainPath, isFile);
     isRun = true;
 //    Loop();
     thrd = std::thread(thrdRun<Server, &Server::Loop>, this);
@@ -67,7 +74,7 @@ void Server::LoopStop()
 
 void Server::Broadcast(const string& data)
 {
-    cout << "send1 " << data << endl;
+//    cout << "send1 " << data << endl;
     auto it = slist.begin(); 
     while(it != slist.end())
     {
@@ -83,7 +90,7 @@ void Server::Broadcast(const string& data)
 
 int Server::Send(int socket, const string& data)
 {
-    cout << "send to " << socket << " " << data << endl;
+//    cout << "send to " << socket << " " << data << endl;
     int len = ::write(socket, data.data(), data.size());
     if(len < 0)
         ::close(socket);
